changed locations of views and urls
