+++ revolv/templates/base/partials/cash_flow_table.html
@@ -1,5 +1,5 @@
 {% comment %}
-	stuff here who knows tbh
+	This partial renders a cash flow statement that is later inserted into the accounting page.
 {% endcomment %}
 
 {% block body %}
